requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ytmusicapi==1.3.0
yt-dlp==2023.11.16
Pillow==10.1.0
requests==2.31.0
pydantic==2.5.0

